+++ js/modules/arrays.js
@@ -1,0 +1,57 @@
+const descriptions = [
+  'desc1',
+  'desc2',
+  'desc3',
+  'desc4',
+  'desc5',
+  'desc6',
+  'desc7',
+  'desc8',
+  'desc9',
+  'desc10',
+  'desc11',
+  'desc12',
+  'desc13',
+  'desc14',
+  'desc15',
+  'desc16',
+  'desc17',
+  'desc18',
+  'desc19',
+  'desc20',
+  'desc21',
+  'desc22',
+];
+const comments = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+const names = [
+  'Ольга',
+  'Святослав',
+  'Глеб',
+  'Максим',
+  'Николай',
+  'Ирина',
+  'Павел',
+  'Виктор',
+  'Алла',
+  'Михаил',
+  'Григорий',
+  'Инна',
+  'Софья',
+  'Марина',
+  'Арсений',
+  'Милена',
+  'Лариса',
+  'Дина',
+  'Валерий',
+  'Руслан',
+];
+export{descriptions, comments, names};
